Segregating ContextMenu into its own control
